Interflux #0 (#503)
* cosmetics, minor fixes

* comments and renames

* Rename Ethereum to ETH

* break InteropSettings into sections

* Rename ETHClientBase to ETHClient

* Use enums instead of ints in ConversionRequest and add destination chain

* renamed directory. EthereumClient => ETHClient

* move IETHClient to the same file as ETHClient

* refactoring

* add TargetChain to deposit

* BuildInterFluxTransaction wallet controller method

* reverted Deposit constructor changes

* Update src/Stratis.Bitcoin.Features.Interop/InteropSettings.cs

+++ src/Stratis.SmartContracts.IntegrationTests/Interop/ConversionRequestTests.cs
@@ -22,9 +22,9 @@ namespace Stratis.SmartContracts.IntegrationTests
             var request = new ConversionRequest()
             {
                 RequestId = "requestId",
-                RequestType = (int)ConversionRequestType.Mint,
+                RequestType = ConversionRequestType.Mint,
                 Processed = false,
-                RequestStatus = (int)ConversionRequestStatus.Unprocessed,
+                RequestStatus = ConversionRequestStatus.Unprocessed,
                 Amount = 100000000,
                 BlockHeight = 123,
                 DestinationAddress = ""
@@ -47,9 +47,9 @@ namespace Stratis.SmartContracts.IntegrationTests
             var request = new ConversionRequest()
             {
                 RequestId = "requestId",
-                RequestType = (int)ConversionRequestType.Mint,
+                RequestType = ConversionRequestType.Mint,
                 Processed = false,
-                RequestStatus = (int)ConversionRequestStatus.Unprocessed,
+                RequestStatus = ConversionRequestStatus.Unprocessed,
                 Amount = 100000000,
                 BlockHeight = 123,
                 DestinationAddress = ""
